Migration of data upload logic from AppDelegate to a separate DataManager class
